+++ R/benchmark_markov.R
@@ -82,7 +82,7 @@ benchmark_markov <- function(markov_model = NULL, reference = NULL,
                          mc.cores = 4
       )
     },
-    "SpeedyMarkov - Rcpp simulation - furrr::future_map (4 cores)" = {
+    "Rcpp simulation - furrr::future_map (4 cores)" = {
       markov_ce_pipeline(markov_model(),
                          duration = duration, 
                          samples = samples,